Updated password reset limit code per review
