+++ features/step_definitions/user_steps.rb
@@ -226,10 +226,10 @@ When /^the user "(.*?)" accepts all co-creator requests$/ do |login|
   user.creatorships.unapproved.each(&:accept!)
 end
 
-When /^I request a password reset for (.*)$/ do |login|
+When "I request a password reset for {string}" do |login|
   step(%{I am on the login page})
   step(%{I follow "Reset password"})
-  step(%{I fill in "Email address or user name" with #{login}})
+  step(%{I fill in "Email address or user name" with "#{login}"})
   step(%{I press "Reset Password"})
 end
